clip-ai/requirements.txt
fastapi
uvicorn
torch
transformers
Pillow
python-multipart
requests
